+++ js/functions.js
@@ -1,6 +1,31 @@
+import { ready, getCustomElement } from 'https://cdn.kernvalley.us/js/std-js/functions.js';
 export function getMarkers() {
 	return Array.from(document.querySelectorAll('leaflet-marker'));
 }
+
+export async function consumeHandler({ files }) {
+	if (files.length === 1) {
+		try {
+			const fileHandle = files[0];
+			const [GeoJSON, file] = await Promise.all([
+				getCustomElement('leaflet-geojson'),
+				fileHandle.getFile(),
+				ready(),
+			]);
+
+			const map = document.querySelector('leaflet-map');
+			const geoJSON = new GeoJSON();
+			geoJSON.data = file;
+			geoJSON.stroke = 'red';
+			geoJSON.weight = 4;
+			map.append(geoJSON);
+		} catch(err) {
+			console.error(err);
+			alert(err.message);
+		}
+	}
+}
+
 
 export function getMarkerTitles() {
 	return getMarkers().map(marker => marker.title);